Search extensions modifications from PR comments
